# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LLM Tool Bridge API",
    "description": "An OpenAI-compatible proxy that adds tool/function calling capabilities to any LLM. This proxy intelligently manages multi-turn conversations, handles tool execution, and provides seamless integration with existing OpenAI SDK clients.",
    "version": "1.0.0",
    "contact": {
      "name": "LLM Tool Bridge Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3500",
      "description": "Main Proxy Server"
    },
    {
      "url": "http://localhost:3501",
      "description": "Admin API Server"
    }
  ],
...
  "paths": {
...
    "/api/admin/targets": {
      "get": {
        "tags": ["Admin"],
        "summary": "List LLM targets",
        "description": "Returns all configured LLM targets. Each target represents a backend LLM service that the proxy can route requests to.",
        "operationId": "listTargets",
        "servers": [{"url": "http://localhost:3501"}],
        "responses": {
          "200": {
            "description": "List of configured targets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LLMTarget"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create LLM target",
        "description": "Creates a new LLM target configuration. The target will be available for routing immediately after creation.",
        "operationId": "createTarget",
        "servers": [{"url": "http://localhost:3501"}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTargetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Target created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMTarget"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/targets/{targetId}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get target details",
        "description": "Returns detailed information about a specific LLM target.",
        "operationId": "getTarget",
        "servers": [{"url": "http://localhost:3501"}],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Target details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMTarget"
                }
              }
            }
          },
          "404": {
            "description": "Target not found"
          }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "Update target",
        "description": "Updates an existing LLM target configuration.",
        "operationId": "updateTarget",
        "servers": [{"url": "http://localhost:3501"}],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTargetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Target updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMTarget"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete target",
        "description": "Removes an LLM target from the configuration.",
        "operationId": "deleteTarget",
        "servers": [{"url": "http://localhost:3501"}],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Target deleted successfully"
          },
          "404": {
            "description": "Target not found"
          }
        }
      }
    },
...
    "/api/admin/models/fetch/{targetId}": {
      "post": {
        "tags": ["Admin"],
        "summary": "Fetch models from target",
        "description": "Fetches available models from a specific LLM target and updates the cache.",
        "operationId": "fetchModelsFromTarget",
        "servers": [{"url": "http://localhost:3501"}],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Models fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Model"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LLMTarget": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier"
          },
          "name": {
            "type": "string",
            "description": "Display name"
          },
          "url": {
            "type": "string",
            "description": "Endpoint URL"
          },
          "model": {
            "type": "string",
            "description": "Default model to use"
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether this target is enabled"
          },
          "priority": {
            "type": "integer",
            "description": "Priority for selection (lower is higher priority)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTargetRequest": {
        "type": "object",
        "required": ["name", "url", "model"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name for the target"
          },
          "url": {
            "type": "string",
            "description": "Full URL to the chat completions endpoint"
          },
          "model": {
            "type": "string",
            "description": "Model identifier"
          },
          "apiKey": {
            "type": "string",
            "description": "API key for authentication (optional)"
          },
          "enabled": {
            "type": "boolean",
            "default": true
          },
          "priority": {
            "type": "integer",
            "default": 1
          }
        }
      },
      "UpdateTargetRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "priority": {
            "type": "integer"
          }
        }
      },
      "Model": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Model identifier"
          },
          "name": {
            "type": "string",
            "description": "Model display name"
          },
          "provider": {
            "type": "string",
            "description": "Provider/target name"
          },
          "capabilities": {
            "$ref": "#/components/schemas/ModelCapabilities"
          }
        }
      },
      "ModelCapabilities": {
        "type": "object",
        "properties": {
          "streaming": {
            "type": "boolean",
            "description": "Supports streaming responses"
          },
          "functionCalling": {
            "type": "boolean",
            "description": "Supports function/tool calling"
          },
          "vision": {
            "type": "boolean",
            "description": "Supports image inputs"
          },
          "maxTokens": {
            "type": "integer",
            "description": "Maximum token limit"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Optional chat API key for secured endpoints"
      }
    }
  }
}